report-engine/requirements.txt
playwright==1.49.1
jinja2==3.1.4
pyyaml==6.0.2
requests==2.32.3
